audio_tools/audio_test_requirements.txt
flask>=2.0.0
elevenlabs>=0.2.24
python-dotenv>=0.19.0
sounddevice>=0.4.5
numpy>=1.20.0 